extract SearchInput and FactionSelector from BottomFilterPanel
BottomFilterPanel was 271 lines handling panel animation, search input,
faction selector with tooltip, and faction chips. Extracts:

- SearchInput: text input with responsive width
- FactionSelector: react-select dropdown, info tooltip, faction chips

BottomFilterPanel now ~115 lines focused on panel animation and layout.

Adds tests/bottomfilterpanel-subcomponents.test.ts (6 tests).
